+++ src/test/java/org/ga4gh/testbed/api/model/PhaseTest.java
@@ -13,19 +13,19 @@ public class PhaseTest {
     public Object[][] getData() {
         return new Object[][] {
             {
-                Integer.valueOf(0),
+                Long.valueOf(0),
                 "sequence",
                 "refget sequence endpoint",
                 LocalDateTime.now(),
                 LocalDateTime.now(),
                 Status.PASS,
                 new Summary() {{
-                    setId(0);
+                    setId(Long.valueOf(0));
                     setPassed(10);
                 }},
                 new ArrayList<TestbedTest>() {{
                     add(new TestbedTest() {{
-                        setId(0);
+                        setId(Long.valueOf(0));
                         setTestName("test_sequence_circular");
                     }});
                 }},
@@ -38,7 +38,7 @@ public class PhaseTest {
     }
 
     @Test(dataProvider = "cases")
-    public void testPhase(Integer id, String phaseName, String phaseDescription,
+    public void testPhase(Long id, String phaseName, String phaseDescription,
         LocalDateTime startTime, LocalDateTime endTime, Status status,
         Summary summary, List<TestbedTest> tests, Report report) {
